+++ CHANGELOG.rst
@@ -12,6 +12,8 @@ New Features
 
 Breaking changes
 ^^^^^^^^^^^^^^^^
+- Switch random number generation for drawing emulations from np.random.seed() to np.random.default_rng()
+  (`#495 <https://github.com/MESMER-group/mesmer/pull/495>`_). By `Victoria Bauer`_.
 - Using Cholesky decomposition for finding covariance localization radius and drawing from the multivariate normal distribution (`#408 <https://github.com/MESMER-group/mesmer/pull/408>`_)
   By `Victoria Bauer`_.
 - The supported versions of some dependencies were changed (`#399 <https://github.com/MESMER-group/mesmer/pull/399>`_, `#405 <https://github.com/MESMER-group/mesmer/pull/405>`_):
@@ -42,9 +44,11 @@ Internal Changes
   By `Mathias Hauser`_.
 - Restore compatibility with pandas version v2.2 and xarray version v2024.02 (`#404 <https://github.com/MESMER-group/mesmer/pull/404>`_).
   By `Mathias Hauser`_.
-- Explicitely include all required dependencies (`#448 <https://github.com/MESMER-group/mesmer/pull/448>`_).
+- Explicitly include all required dependencies (`#448 <https://github.com/MESMER-group/mesmer/pull/448>`_).
 - Unshallow the mesmer git repository on rtd (`#456 <https://github.com/MESMER-group/mesmer/pull/456>`_).
   By `Victoria Bauer`_.
+- Use ruff instead of isort and flake8 to lint the code base (`#490 <https://github.com/MESMER-group/mesmer/pull/490>`_).
+  By `Mathias Hauser`_.
 
 Integration of MESMER-M
 ^^^^^^^^^^^^^^^^^^^^^^^
@@ -52,19 +56,32 @@ Integration of MESMER-M
 This release integrates MESMER-M into the existing MESMER infrastructure. This includes
 some refactoring, bugfixes and enhancements of the MESMER-M functionality. Note
 that this led to some numerical changes compared to the MESMER-M publication
-(`Nath et al. 2022 <https://doi.org/10.5194/esd-13-851-2022>`_).
+(Nath et al., `2022 <https://doi.org/10.5194/esd-13-851-2022>`_).
 
 - move MESMER-M scripts into mesmer (
   `#419 <https://github.com/MESMER-group/mesmer/pull/419>`_, and
   `#421 <https://github.com/MESMER-group/mesmer/pull/421>`_).
+- move the harmonic model and power transformer functionalities to the stats module (
+  `#484 <https://github.com/MESMER-group/mesmer/pull/484>`_).
+- add example script for MESMER-M workflow (`#491 <https://github.com/MESMER-group/mesmer/pull/491>`_)
 
-**Auto-Regression**
+Auto-Regression
+~~~~~~~~~~~~~~~
 
 - Implement functions performing the monthly (cyclo-stationary) auto-regression and adapt these functions to
   work with xarray. This includes extracting the drawing of spatially correlated innovations to a
   stand-alone function. (`#473 <https://github.com/MESMER-group/mesmer/pull/473>`_)
+- Remove the bounds of -1 and 1 on the slope of the cyclo-stationary AR(1) process. This bound is not necessary
+  since cyclo-stationarity is also given if the slopes of a few months are (slightly) larger than one. We
+  now return the residuals of the cyclo-stationary AR(1) process to fit the covariance matrix on these residuals.
+  As a consequence, adjustment of the covariance matrix with the AR slopes is no longer necessary.
+  After this, no adjustment is necessary anymore. (`#480 <https://github.com/MESMER-group/mesmer/pull/480>`_)
+  Compare discussion in `#472 <https://github.com/MESMER-group/mesmer/issues/472>`_.
+- Implement function to localize the empirical covarince matrix for each month individually to use in drawing
+  of spatially correlated noise in the AR process. (`#479 <https://github.com/MESMER-group/mesmer/pull/479>`_)
 
-**Yeo-Johnson power transformer**
+Yeo-Johnson power transformer
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 -  Ensure the power transformer yields the correct normalization for more cases (
    `#440 <https://github.com/MESMER-group/mesmer/issues/440>`_):
@@ -74,10 +91,12 @@ that this led to some numerical changes compared to the MESMER-M publication
       `#446 <https://github.com/MESMER-group/mesmer/pull/446>`_)
    -  remove jacobian ``rosen_der`` from fit (
       `#447 <https://github.com/MESMER-group/mesmer/pull/447>`_)
-   -  change opimization method from *SLSQP* to *Nelder-Mead* (
+   -  change optimization method from *SLSQP* to *Nelder-Mead* (
       `#455 <https://github.com/MESMER-group/mesmer/pull/455>`_)
 -  adjust the first guess to assume the data is normally distributed (
    `#429 <https://github.com/MESMER-group/mesmer/pull/429>`_)
+-  make (back-) transformations more stable by using `np.expm1` and `np.log1p`
+   (`#494 <https://github.com/MESMER-group/mesmer/pull/494>`_)
 -  rewrite power transformer to work with xarray, and refactor from a class structure to functions (
    `#442 <https://github.com/MESMER-group/mesmer/pull/442>`_, and
    `#474 <https://github.com/MESMER-group/mesmer/pull/474>`_)
@@ -90,17 +109,21 @@ that this led to some numerical changes compared to the MESMER-M publication
 - add tests (`#430 <https://github.com/MESMER-group/mesmer/pull/430>`_)
 
 
-**Harmonic model**
+Harmonic model
+~~~~~~~~~~~~~~
 
 -  Performance and other optimizations:
 
    - only fit orders up to local minimum and use coeffs from precious order as first guess (`#443 <https://github.com/MESMER-group/mesmer/pull/443>`_)
    - infer the harmonic model order from the coefficients (`#434 <https://github.com/MESMER-group/mesmer/pull/434>`_)
 -  return residuals instead of the loss for the optimization (`#460 <https://github.com/MESMER-group/mesmer/pull/460>`_)
+-  remove fitting of linear regression with yearly temperature (`#415 <https://github.com/MESMER-group/mesmer/pull/415/>_` and
+   `#488 <https://github.com/MESMER-group/mesmer/pull/488>`_) in line with (`Nath et al. 2022 <https://doi.org/10.5194/esd-13-851-2022>`_).
 -  add helper function to upsample yearly data to monthly resolution (
    `#418 <https://github.com/MESMER-group/mesmer/pull/418>`_, and
    `#435 <https://github.com/MESMER-group/mesmer/pull/435>`_)
 - de-duplicate the expression of months in their harmonic form (`#415 <https://github.com/MESMER-group/mesmer/pull/415>`_)
+  move creation of the month array to the deepest level (`#487 <https://github.com/MESMER-group/mesmer/pull/487>`_).
 - fix indexing of harmonic model coefficients (`#415 <https://github.com/MESMER-group/mesmer/pull/415>`_)
 -  Refactor variable names, small code improvements, fixes and clean docstring (
    `#415 <https://github.com/MESMER-group/mesmer/pull/415>`_,
@@ -109,6 +132,7 @@ that this led to some numerical changes compared to the MESMER-M publication
 - add tests (
   `#431 <https://github.com/MESMER-group/mesmer/pull/431>`_, and
   `#458 <https://github.com/MESMER-group/mesmer/pull/458>`_)
+- add function to generate fourier series using xarray (`#478 <https://github.com/MESMER-group/mesmer/pull/478>`_)
 
 By `Victoria Bauer`_ and `Mathias Hauser`_.